Shape3D fixed

+++ src/Chapter9/Shape3D.java
@@ -12,5 +12,5 @@ package Chapter9;
 public interface Shape3D {
     public double getVolume();
     public double getSurfaceArea();
-    public double getSides();
+    public int getSides();
 }
